# Stop checking uniqueKeys once we've found 10 errors