This includes the benchmark test cases from https://github.com/go-yaml/yaml/pull/543.