@cjcullen 2019-11-08
It's kinda overkill to "benchmark" this change, but:
```
$ benchcmp old.txt new.txt
benchmark                    old ns/op      new ns/op     delta
BenchmarkDuplicateKeys-6     2370532130     6586106       -99.72%

benchmark                    old allocs     new allocs     delta
BenchmarkDuplicateKeys-6     2002187        8139           -99.59%

benchmark                    old bytes     new bytes     delta
BenchmarkDuplicateKeys-6     171812904     1012729       -99.41%
```